pyproject.toml
[project]
name = "e"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "openai>=1.76.0",
    "pillow>=11.2.1",
    "polars>=1.28.0",
    "pydantic>=2.11.3",
    "python-magic>=0.4.27",
    "ruff>=0.11.7",
    "sqlmodel>=0.0.24",
    "tqdm>=4.67.1",
    "wandb[media]>=0.19.10",
]
